requirements.txt
tensorflow
numpy 
scikit-learn
matplotlib
st-annotated-text
streamlit
opencv-python